django-backend/requirements.txt
# Django Backend for CLAP - Requirements
# Python 3.10+

# Core Django
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8

# Database
psycopg2-binary==2.9.9

# Authentication & Security
bcrypt==4.1.2

# OpenAI Integration
openai==1.10.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
django-extensions==3.2.3
ipython==8.20.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0
